添加mapping字段dto

+++ datapool-common/src/main/java/cn/ict/jwdsj/datapool/common/entity/indexmanage/MappingColumn.java
@@ -42,6 +42,6 @@ public class MappingColumn extends BaseEntity {
     private boolean displayed = true;
 
     @Column(name = "boost", nullable = false)
-    private double boost = 1;
+    private float boost = 1;
 
 }
